requirements.txt
aiogram==2.25.1
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
Babel==2.9.1
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
frozenlist==1.4.1
idna==3.6
itsdangerous==2.2.0
Jinja2==3.1.3
magic-filter==1.0.12
MarkupSafe==2.1.5
multidict==6.0.5
peewee==3.17.1
python-decouple==3.8
pytz==2024.1
Werkzeug==3.0.2
yarl==1.9.4
